+++ jfr-streaming/src/main/java/io/opentelemetry/contrib/jfr/metrics/memory/PerThreadObjectAllocationInNewTLABHandler.java
@@ -1,14 +1,12 @@
-package org.jfr.metrics.memory;
+package io.opentelemetry.contrib.jfr.metrics.memory;
 
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.metrics.BoundDoubleHistogram;
 import io.opentelemetry.api.metrics.Meter;
+import io.opentelemetry.contrib.jfr.Constants;
+import io.opentelemetry.contrib.jfr.metrics.RecordedEventHandler;
 import jdk.jfr.consumer.RecordedEvent;
-import org.jfr.metrics.RecordedEventHandler;
-
-import static org.jfr.metrics.Constants.KILOBYTES;
-import static org.jfr.metrics.Constants.THREAD_NAME;
 
 /** This class aggregates all TLAB allocation JFR events for a single thread */
 public final class PerThreadObjectAllocationInNewTLABHandler implements RecordedEventHandler {
@@ -28,10 +26,10 @@ public final class PerThreadObjectAllocationInNewTLABHandler implements Recorded
   }
 
   public PerThreadObjectAllocationInNewTLABHandler init() {
-    var attr = Attributes.of(AttributeKey.stringKey(THREAD_NAME), threadName);
+    var attr = Attributes.of(AttributeKey.stringKey(Constants.THREAD_NAME), threadName);
     var builder = otelMeter.histogramBuilder(HISTOGRAM_NAME);
     builder.setDescription(DESCRIPTION);
-    builder.setUnit(KILOBYTES);
+    builder.setUnit(Constants.KILOBYTES);
     histogram = builder.build().bind(attr);
     return this;
   }
