+++ jfr-streaming/src/main/java/io/opentelemetry/contrib/jfr/metrics/network/NetworkReadHandler.java
@@ -1,9 +1,9 @@
-package org.jfr.metrics.network;
+package io.opentelemetry.contrib.jfr.metrics.network;
 
 import io.opentelemetry.api.metrics.Meter;
-import org.jfr.metrics.AbstractThreadDispatchingHandler;
-import org.jfr.metrics.RecordedEventHandler;
-import org.jfr.metrics.ThreadGrouper;
+import io.opentelemetry.contrib.jfr.metrics.AbstractThreadDispatchingHandler;
+import io.opentelemetry.contrib.jfr.metrics.RecordedEventHandler;
+import io.opentelemetry.contrib.jfr.metrics.ThreadGrouper;
 
 public class NetworkReadHandler extends AbstractThreadDispatchingHandler {
   public static final String EVENT_NAME = "jdk.SocketRead";
