+++ jfr-streaming/src/main/java/io/opentelemetry/contrib/jfr/metrics/network/PerThreadNetworkReadHandler.java
@@ -1,13 +1,12 @@
-package org.jfr.metrics.network;
+package io.opentelemetry.contrib.jfr.metrics.network;
 
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.Attributes;
 import io.opentelemetry.api.metrics.BoundDoubleHistogram;
 import io.opentelemetry.api.metrics.Meter;
+import io.opentelemetry.contrib.jfr.Constants;
+import io.opentelemetry.contrib.jfr.metrics.RecordedEventHandler;
 import jdk.jfr.consumer.RecordedEvent;
-import org.jfr.metrics.RecordedEventHandler;
-
-import static org.jfr.metrics.Constants.*;
 
 public class PerThreadNetworkReadHandler implements RecordedEventHandler {
   private static final String DESCRIPTION_BYTES = "Bytes Read";
@@ -27,16 +26,16 @@ public class PerThreadNetworkReadHandler implements RecordedEventHandler {
   }
 
   public PerThreadNetworkReadHandler init() {
-    var attr = Attributes.of(AttributeKey.stringKey(THREAD_NAME), threadName);
+    var attr = Attributes.of(AttributeKey.stringKey(Constants.THREAD_NAME), threadName);
 
     var builder = otelMeter.histogramBuilder(JFR_SOCKET_READ_BYTES_READ);
     builder.setDescription(DESCRIPTION_BYTES);
-    builder.setUnit(KILOBYTES);
+    builder.setUnit(Constants.KILOBYTES);
     bytesHistogram = builder.build().bind(attr);
 
     builder = otelMeter.histogramBuilder(JFR_SOCKET_READ_DURATION);
     builder.setDescription(DESCRIPTION_DURATION);
-    builder.setUnit(MILLISECONDS);
+    builder.setUnit(Constants.MILLISECONDS);
     durationHistogram = builder.build().bind(attr);
 
     return this;
